GetTotal price type int to decimal and it returns the sum of linetotal

+++ Checkout/Checkout.BLL/Interfaces/ICheckoutManager.cs
@@ -12,6 +12,6 @@ namespace Checkout.BLL.Interfaces
         void Scan(string sku);
         List<ShopingCart> GetItems();
         ShopingCart GetItem(string sku);
-        int GetTotalPrice();
+        decimal GetTotalPrice();
     }
 }
